uc-x/requirements.txt
google-genai
python-dotenv
